incorporate cropping/zooming on images
a lot of detail is being lost by resizing images from 1200+ px
resolution to 224. This will now randomly grab smaller portions of
images, say 400x400, to then resize to 224x224. This essentially
splits each image into random zoomed windows, which hopefully
provides more detail to the classifier.

This is possible because the fields generally cover most of the field
view, so a random zoom from the left size will usually, but not always,
be extremely similar to  random zoom from the right size. The part of
the image is usually sky, this also limits the zooming to the bottom
50% of the image.